commiting latest

+++ evaluate.py
@@ -1,3 +1,9 @@
+#torch base function package
+import torch 
+#torch nn class definition. We will inherit Module class from this
+#this will help do the leg work of tracking and updating grads. Also 
+#helpps in setting
+import torch.nn as nn 
 from train import * 
 
 def evaluate(model, iterator, criterion):
@@ -21,7 +27,7 @@ def evaluate(model, iterator, criterion):
             predictions = model(text, text_lengths).squeeze()
             
             #compute loss and accuracy
-            loss = criterion(predictions, batch.icmteam)
+            loss = criterion(predictions, batch.icmteam.long())
             acc = multi_acc(predictions, batch.icmteam)
             
             #keep track of loss and accuracy
